+++ frontend/src/app/main/ui/dashboard/comments.cljs
@@ -32,7 +32,7 @@
         show-dropdown  (mf/use-fn #(reset! show-dropdown? true))
         hide-dropdown  (mf/use-fn #(reset! show-dropdown? false))
         threads-map    (mf/deref refs/comment-threads)
-        users          (mf/deref refs/current-file-comments-users)
+        users          (mf/deref refs/current-team-comments-users)
 
         tgroups        (->> (vals threads-map)
                             (sort-by :modified-at)